split test & SUP files in packages set maven version in pom.xml
