+++ src/main/java/hwr/oop/todo/application/Task.java
@@ -1,4 +1,4 @@
-package hwr.oop.todo;
+package hwr.oop.todo.application;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
