+++ src/main/java/hwr/oop/todo/application/TaskPriority.java
@@ -1,4 +1,4 @@
-package hwr.oop.todo;
+package hwr.oop.todo.application;
 
 public enum TaskPriority {
     HIGH,
