+++ src/main/java/hwr/oop/todo/application/ToDoList.java
@@ -1,4 +1,4 @@
-package hwr.oop.todo;
+package hwr.oop.todo.application;
 
 import java.security.InvalidParameterException;
 import java.util.ArrayList;
