+++ src/main/java/hwr/oop/todo/persistence/csv/CSVReader.java
@@ -1,4 +1,6 @@
-package hwr.oop.todo;
+package hwr.oop.todo.persistence.csv;
+
+import hwr.oop.todo.application.*;
 
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
