+++ src/main/java/hwr/oop/todo/persistence/csv/CSVUpdate.java
@@ -1,4 +1,8 @@
-package hwr.oop.todo;
+package hwr.oop.todo.persistence.csv;
+
+import hwr.oop.todo.application.Project;
+import hwr.oop.todo.application.Task;
+import hwr.oop.todo.application.ToDoList;
 
 import java.io.*;
