+++ src/main/java/hwr/oop/todo/ui/cli/MainMenu.java
@@ -1,4 +1,4 @@
-package hwr.oop.todo;
+package hwr.oop.todo.ui.cli;
 
 import java.io.InputStream;
 import java.io.OutputStream;
