+++ src/main/java/hwr/oop/todo/ui/cli/ManageToDoMenu.java
@@ -1,4 +1,8 @@
-package hwr.oop.todo;
+package hwr.oop.todo.ui.cli;
+
+import hwr.oop.todo.ui.Main;
+import hwr.oop.todo.application.ToDoList;
+import hwr.oop.todo.persistence.csv.CSVReader;
 
 import java.io.InputStream;
 import java.io.OutputStream;
