+++ src/main/java/hwr/oop/todo/ui/cli/MenuInputHandler.java
@@ -1,4 +1,4 @@
-package hwr.oop.todo;
+package hwr.oop.todo.ui.cli;
 
 import java.io.PrintStream;
 import java.util.HashMap;
