+++ src/main/java/hwr/oop/todo/ui/cli/ProjectMenu.java
@@ -1,4 +1,7 @@
-package hwr.oop.todo;
+package hwr.oop.todo.ui.cli;
+
+import hwr.oop.todo.ui.Main;
+import hwr.oop.todo.application.Project;
 
 import java.io.InputStream;
 import java.io.OutputStream;
