+++ src/main/java/hwr/oop/todo/ui/cli/TodoMainMenu.java
@@ -1,4 +1,8 @@
-package hwr.oop.todo;
+package hwr.oop.todo.ui.cli;
+
+import hwr.oop.todo.ui.Main;
+import hwr.oop.todo.application.Task;
+import hwr.oop.todo.persistence.csv.CSVCreate;
 
 import java.io.IOException;
 import java.io.InputStream;
