+++ src/test/java/hwr/oop/todo/application/TaskTagTest.java
@@ -1,4 +1,4 @@
-package hwr.oop.todo;
+package hwr.oop.todo.application;
 
 import org.junit.jupiter.api.Test;
